+++ src/bundler/rules/url.ts
@@ -1,9 +1,19 @@
 import { STATIC_ASSETS } from '../../config/constants';
 
+const imageInlineSizeLimit = parseInt(
+  process.env.IMAGE_INLINE_SIZE_LIMIT || '10000'
+);
+
 function urlLoader(): object[] {
   const PATHNAME = `${BalmJS.config.paths.target.js}/${STATIC_ASSETS}/`;
   const FILENAME = '[name].[hash:8].[ext]';
-  const limit: number = BalmJS.config.scripts.base64Limit;
+  const options = Object.assign(
+    {
+      limit: imageInlineSizeLimit, // Loads files as `base64` encoded URL
+      name: BalmJS.file.assetsPath(`${PATHNAME}${FILENAME}`)
+    },
+    BalmJS.config.scripts.urlLoaderOptions
+  );
 
   return [
     // "url" loader works like "file" loader except that it embeds assets
@@ -12,26 +22,17 @@ function urlLoader(): object[] {
     {
       test: /\.(bmp|gif|jpe?g|png|svg)(\?.*)?$/,
       loader: 'url-loader',
-      options: {
-        limit,
-        name: BalmJS.file.assetsPath(`${PATHNAME}${FILENAME}`)
-      }
+      options
     },
     {
       test: /\.(mp4|webm|ogg|mp3|wav|flac|aac)(\?.*)?$/,
       loader: 'url-loader',
-      options: {
-        limit,
-        name: BalmJS.file.assetsPath(`${PATHNAME}${FILENAME}`)
-      }
+      options
     },
     {
       test: /\.(woff2?|eot|ttf|otf)(\?.*)?$/,
       loader: 'url-loader',
-      options: {
-        limit,
-        name: BalmJS.file.assetsPath(`${PATHNAME}${FILENAME}`)
-      }
+      options
     }
   ];
 }