+++ test-workspace/vue-spa/scripts/views/layouts/app.vue
@@ -1,6 +1,7 @@
 <template>
   <div class="container">
     <img :src="logo" alt />
+    <img :src="require('@/assets/logo.png')" alt />
     <hello></hello>
   </div>
 </template>